refactor(frontend): convention cleanup — extract interfaces, suggestions, chart constants
Three behavior-preserving cleanups to match the frontend standards:
- Inline interfaces/types declared in component and service .ts files moved into
  per-domain *.model.ts files (analyze-request, prepare-request, job-filters,
  seniority-level, login-response, recent-calls-filters, llm-provider, etc.) and
  importers updated.
- Hardcoded LLM provider/model suggestions extracted from admin-llm-settings into
  a dedicated constants module.
- Unnamed numeric literals in the analytics charts (trend-line, salary-band,
  funnel-chart) replaced with named constants.

Closes #56
