+++ frontend/src/app/core/models/login-response.model.ts
@@ -1,0 +1,4 @@
+export interface LoginResponse {
+  access_token: string;
+  token_type: string;
+}
