+++ frontend/src/app/core/services/matching.service.ts
@@ -2,18 +2,10 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { environment } from '../../../environments/environment';
+import { AnalyzeRequest } from '../../pages/matching/models/analyze-request.model';
 import { EvaluateRequest } from '../../pages/matching/models/evaluate-request.model';
 import { EvaluationResult } from '../../pages/matching/models/evaluation-result.model';
 import { MatchResult } from '../../pages/matching/models/match-result.model';
-
-export interface AnalyzeRequest {
-  job_id: string;
-  cv_id: string;
-  job_description: string;
-  job_skills: string[];
-  cv_summary: string;
-  cv_skills: string[];
-}
 
 @Injectable({ providedIn: 'root' })
 export class MatchingService {
