+++ frontend/src/app/core/services/research.service.ts
@@ -3,11 +3,7 @@ import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { environment } from '../../../environments/environment';
 import { CompanyResearch } from '../../pages/tracking/models/company-research.model';
-
-export interface ResearchRequest {
-  company_name: string;
-  job_description: string;
-}
+import { ResearchRequest } from '../../pages/tracking/models/research-request.model';
 
 @Injectable({ providedIn: 'root' })
 export class ResearchService {
