+++ frontend/src/app/pages/admin/models/extra-param.model.ts
@@ -1,0 +1,4 @@
+export interface ExtraParam {
+  key: string;
+  value: string;
+}
