+++ frontend/src/app/pages/admin/models/llm-provider.model.ts
@@ -1,0 +1,3 @@
+import { LLM_PROVIDERS } from '../constants/llm-provider-suggestions';
+
+export type LLMProvider = (typeof LLM_PROVIDERS)[number];
