+++ frontend/src/app/pages/admin/models/recent-calls-filters.model.ts
@@ -1,0 +1,8 @@
+export interface RecentCallsFilters {
+  limit?: number;
+  offset?: number;
+  provider?: string;
+  model?: string;
+  feature_key?: string;
+  days?: number;
+}
