+++ frontend/src/app/pages/analytics/analytics.component.ts
@@ -5,7 +5,8 @@ import { FunnelMetrics } from './models/funnel-metrics.model';
 import { MarketIntel } from './models/market-intel.model';
 import { SkillGap } from './models/skill-gap.model';
 import { TargetSalary } from './models/target-salary.model';
-import { BarChartComponent, BarRow } from './components/bar-chart/bar-chart.component';
+import { BarRow } from './models/bar-row.model';
+import { BarChartComponent } from './components/bar-chart/bar-chart.component';
 import { FunnelChartComponent } from './components/funnel-chart/funnel-chart.component';
 import { TrendLineComponent } from './components/trend-line/trend-line.component';
 import { SalaryBandComponent } from './components/salary-band/salary-band.component';
