+++ frontend/src/app/pages/analytics/models/bar-row.model.ts
@@ -1,0 +1,6 @@
+export interface BarRow {
+  label: string;
+  value: number;
+  pct: number;
+  note?: string;
+}
