+++ frontend/src/app/pages/applications/models/cover-letter-tone.model.ts
@@ -1,0 +1,1 @@
+export type CoverLetterTone = 'professional' | 'enthusiastic' | 'concise';
