+++ frontend/src/app/pages/ingestion/models/seniority-level.model.ts
@@ -1,0 +1,1 @@
+export type SeniorityLevel = 'intern' | 'junior' | 'mid' | 'senior' | 'lead' | 'unknown';
