+++ frontend/src/app/pages/interview/models/prepare-request.model.ts
@@ -1,0 +1,5 @@
+export interface PrepareRequest {
+  job_title: string;
+  company: string;
+  description: string;
+}
