+++ frontend/src/app/pages/matching/models/analyze-request.model.ts
@@ -1,0 +1,8 @@
+export interface AnalyzeRequest {
+  job_id: string;
+  cv_id: string;
+  job_description: string;
+  job_skills: string[];
+  cv_summary: string;
+  cv_skills: string[];
+}
