+++ frontend/src/app/pages/profile/components/manual-fields-form/manual-fields-form.component.ts
@@ -8,22 +8,11 @@ import {
 } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CandidateProfile } from '../../models/candidate-profile.model';
-import {
-  ProfileManualFieldsUpdate,
-  ProfileService,
-} from '../../../../core/services/profile.service';
+import { ManualFieldsFormState } from '../../models/manual-fields-form-state.model';
+import { ProfileManualFieldsUpdate } from '../../models/profile-manual-fields-update.model';
+import { ProfileService } from '../../../../core/services/profile.service';
 
-type FormState = {
-  name: string;
-  email: string;
-  phone: string;
-  location: string;
-  linkedin_url: string;
-  github_url: string;
-  portfolio_url: string;
-};
-
-const FIELDS: ReadonlyArray<keyof FormState> = [
+const FIELDS: ReadonlyArray<keyof ManualFieldsFormState> = [
   'name',
   'email',
   'phone',
@@ -33,7 +22,7 @@ const FIELDS: ReadonlyArray<keyof FormState> = [
   'portfolio_url',
 ];
 
-function snapshot(profile: CandidateProfile): FormState {
+function snapshot(profile: CandidateProfile): ManualFieldsFormState {
   return {
     name: profile.name ?? '',
     email: profile.email ?? '',
@@ -58,7 +47,7 @@ export class ManualFieldsFormComponent {
 
   profile = input.required<CandidateProfile>();
 
-  form = signal<FormState>({
+  form = signal<ManualFieldsFormState>({
     name: '',
     email: '',
     phone: '',
@@ -67,7 +56,7 @@ export class ManualFieldsFormComponent {
     github_url: '',
     portfolio_url: '',
   });
-  baseline = signal<FormState>(this.form());
+  baseline = signal<ManualFieldsFormState>(this.form());
   saving = signal(false);
   error = signal('');
   savedFlash = signal(false);
@@ -80,7 +69,7 @@ export class ManualFieldsFormComponent {
     });
   }
 
-  set<K extends keyof FormState>(key: K, value: string): void {
+  set<K extends keyof ManualFieldsFormState>(key: K, value: string): void {
     this.form.update((current) => ({ ...current, [key]: value }));
   }
