+++ frontend/src/app/pages/profile/models/cover-letter-template-editing-state.model.ts
@@ -1,0 +1,4 @@
+export type CoverLetterTemplateEditingState =
+  | { mode: 'closed' }
+  | { mode: 'new' }
+  | { mode: 'edit'; id: string };
