+++ frontend/src/app/pages/profile/models/cover-letter-template-upsert.model.ts
@@ -1,0 +1,8 @@
+export interface CoverLetterTemplateUpsert {
+  name?: string;
+  tone?: string;
+  language?: string;
+  opening?: string;
+  body?: string;
+  signature?: string;
+}
