+++ frontend/src/app/pages/profile/models/manual-fields-form-state.model.ts
@@ -1,0 +1,9 @@
+export interface ManualFieldsFormState {
+  name: string;
+  email: string;
+  phone: string;
+  location: string;
+  linkedin_url: string;
+  github_url: string;
+  portfolio_url: string;
+}
