+++ frontend/src/app/pages/profile/models/profile-manual-fields-update.model.ts
@@ -1,0 +1,9 @@
+export interface ProfileManualFieldsUpdate {
+  name?: string | null;
+  email?: string | null;
+  phone?: string | null;
+  location?: string | null;
+  linkedin_url?: string | null;
+  github_url?: string | null;
+  portfolio_url?: string | null;
+}
