+++ frontend/src/app/pages/profile/models/upload-cv-request.model.ts
@@ -1,0 +1,4 @@
+export interface UploadCVRequest {
+  tex_content: string;
+  language: string;
+}
