+++ frontend/src/app/pages/tracking/models/research-request.model.ts
@@ -1,0 +1,4 @@
+export interface ResearchRequest {
+  company_name: string;
+  job_description: string;
+}
